build.gradle se añade el classpath de navigation safe args build.gradle app se añaden los plugings de navigation saveargs y parcelize AndroidManifest se agrega el usesClearTestTraffic en true Extension se agrega la funcion de extension ImageView.load2 invocando a la libreria picasso Menu Fragment se agrega los args del navArgs se setea la bandera segun el argumento site.id se cambia la forma de invocar al navigate de manera segura mediant el frameDirectios.action en el producto model se cambia el tipo de dato de la variable price de entero a flotante ProductSearchFragment se lee el args desde el navArgs y se añade un progressBar mientras realiza la busqueda se añade el parametro siteId para el getProductQuery la clase SiteData se vuelve @Parcelable se usa la forma segura de invocar un nuevo fragmente dcon el FragmenteDIrections

+++ app/src/main/java/com/sestepa/melisearch/entities/product/data/remote/ProductModel.kt
@@ -8,7 +8,7 @@ data class ProductModel(
 		@SerializedName("id") val id: String,
 		@SerializedName("site_id") val siteId: String,
 		@SerializedName("title") val title: String,
-		@SerializedName("price") val price: Int,
+		@SerializedName("price") val price: Float,
 		@SerializedName("currency_id") val currencyId: String,
 		@SerializedName("available_quantity") val availableQuantity: Int,
 		@SerializedName("sold_quantity") val soldQuantity: Int,
